media-source/MediaSource-algorithm-DurationChange.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Check "duration change" algorithm of MediaSource Object</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = '/media/white.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vp8"';

function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(new Uint8Array(xhr.response));
    };
}
//check the browser supports the MIME used in this test
function canPlayType(mime) {
    if(video.canPlayType(mime) != 'probably') {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test.");
        });
        this.done();
        return false;
    }
    return true;
}
async_test(function(t) {
    if(!canPlayType.bind(t)(mime)) {
        return;
    }
    GET(FILE, function(data) {
        var mediaSource = new MediaSource();
        video.controller = new MediaController();
        var controller_duration = 0;
        var mediaSource_duration = 0;
        var durationchangeNum = 0;
        mediaSource.addEventListener('sourceopen', function(e){
            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            sourceBuffer.addEventListener('updateend', function(e) {
                mediaSource.duration = 5;//change the duration
            });
            sourceBuffer.appendBuffer(data);
        });
        video.addEventListener('durationchange', function(e) {
            console.log('durationchange');
            durationchangeNum++;
            controller_duration = video.controller.duration;
            mediaSource_duration = mediaSource.duration;

            if(durationchangeNum > 1) {
                t.step(function() {
                    assert_equals(durationchangeNum, 2, 'One of two is fired by video element, another one is fired by MediaSource.');
                    assert_equals(controller_duration, 5);
                    assert_equals(mediaSource_duration, 5);
                });
                t.done();
            }
        });
        video.src = window.URL.createObjectURL(mediaSource);
    });
}, 'Check "duration change" algorithm of MediaSource Object');

</script>
</body>
</html>
